requirements.txt
aiogram
sqlalchemy
python-dotenv
asyncpg
aiosqlite
environs